improved left naviagtion with proper components distribution and made animations work again
