+++ src/index.js
@@ -1,6 +1,7 @@
 import React,{Component} from 'react';
-import ReactDOM from 'react-dom'
-import LeftNav from './components/leftNav'
+import ReactDOM from 'react-dom';
+import ReactWOW from 'react-wow';
+import LeftNav from './components/leftComponents/leftNav'
 import About_me from './components/about-me.js';
 import Experience from './components/experience.js';
 import Education from './components/education.js';
